+++ src/components/Datepicker/Datepicker.jsx
@@ -9,13 +9,8 @@ import PropTypes from "prop-types";
 import Tappable from "react-tapper";
 import { classNames } from "../../utils";
 import styles from "./styles.css";
-import {
-  getYearArray,
-  validValues,
-  mapToArray,
-  getYearsByNum,
-  getDID
-} from "./helpers";
+import DatePickerPad from "./DatepickerPad.jsx";
+import { getYearArray, validValues, getYearsByNum, getDID } from "./helpers";
 
 const DatePicker = ({
   range,
@@ -47,7 +42,6 @@ const DatePicker = ({
   const [showedState, setShowedState] = useState(show);
   const [yearIndexesState] = useState(yearIndexes);
   const [pads, setPads] = useState([]);
-  const [popupClass, setPopupClass] = useState("");
   const isRange = valuesState.length > 1;
 
   useEffect(() => {
@@ -70,201 +64,40 @@ const DatePicker = ({
     };
   }, []);
 
-  const renderPad = () => {
-    if (isRange > 1) {
-      setPads([optionPad(0), optionPad(1)]);
-      setPopupClass("range");
-    } else {
-      setPads([optionPad(0)]);
-    }
-  };
-
-  const optionPad = (padIndex) => {
-    const values = valuesState;
-    const value = values[padIndex];
+  const padByIndex = (padIndex) => {
     const labelYears = labelYearsState;
     const labelMonths = labelMonthsState;
-    console.log({
-      yearsState,
-      labelYearsState,
-      yearIndexesState,
-      labelMonthsState,
-      valuesState
-    });
     const labelYear = (labelYears[padIndex] =
       labelYears[padIndex] || value.year);
     const labelMonth = (labelMonths[padIndex] =
       labelMonths[padIndex] || value.month);
-    const ymArr = yearsState;
-    const months = Array.isArray(lang)
-      ? lang
-      : Array.isArray(lang.months)
-      ? lang.months
-      : [];
-    let prevCss = "";
-    let prevMonthCss = "";
-    let nextCss = "";
-    let nextMonthCss = "";
-    const yearMaxIdx = ymArr.length - 1;
-    const yearIdx = yearIndexesState[padIndex]; // yearMaxIdx
-
-    if (yearIdx === 0) prevCss = "disable";
-    if (yearIdx === yearMaxIdx) nextCss = "disable";
-
-    const yearActive = labelYear === value.year;
-    const atMinYear = labelYear === ymArr[0].year;
-    const atMaxYear = labelYear === ymArr[yearMaxIdx].year;
-    let otherValue = false;
-    if (values.length > 1) {
-      otherValue = values[1 - padIndex];
-    }
-
-    const labelTextKey = padIndex === 0 ? "from" : "to";
-    let labelPreText;
-    if (otherValue && lang[labelTextKey]) {
-      labelPreText = <b>{lang[labelTextKey]}</b>;
-    }
-
-    if (labelMonth === 1 || (atMinYear && labelMonth === ymArr[0].min))
-      prevMonthCss = "disable";
-    if (
-      labelMonth === 12 ||
-      (atMaxYear && labelMonth === ymArr[yearMaxIdx].max)
-    )
-      nextMonthCss = "disable";
-
     return (
-      <div className={styles.rmpPad} key={padIndex}>
-        <div>
-          <label>
-            {labelPreText}
-            {labelYear}
-          </label>
-          <i
-            className={classNames(styles.rmpBtn, styles.prev, styles[prevCss])}
-            data-id={padIndex}
-            onClick={handlePrevYearClick}>
-            {"<"}
-          </i>
-          <i
-            className={classNames(styles.rmpBtn, styles.next, styles[nextCss])}
-            data-id={padIndex}
-            onClick={handleNextYearClick}>
-            {">"}
-          </i>
-        </div>
-        {!isMonthsMode && (
-          <div>
-            <label>{labelMonth}</label>
-            <i
-              className={classNames(
-                styles.rmpBtn,
-                styles.prevMnth,
-                styles[prevMonthCss]
-              )}
-              data-id={padIndex}
-              onClick={
-                prevMonthCss !== "disable" ? handlePrevMonthClick : undefined
-              }>
-              {"<"}
-            </i>
-            <i
-              className={classNames(
-                styles.rmpBtn,
-                styles.nextMonth,
-                styles[nextMonthCss]
-              )}
-              data-id={padIndex}
-              onClick={
-                nextMonthCss !== "disable" ? handleNextMonthClick : undefined
-              }>
-              {">"}
-            </i>
-          </div>
-        )}
-        <ul>
-          {isMonthsMode &&
-            mapToArray(12, (i) => {
-              let css = "";
-              const m = i + 1;
-              if (yearActive && m === value.month) {
-                css = "active";
-              }
-              if (
-                values.length > 1 &&
-                padIndex === 0 &&
-                (labelYear > value.year ||
-                  (labelYear === value.year && m > value.month))
-              ) {
-                css = "select";
-              }
-              if (
-                values.length > 1 &&
-                padIndex === 1 &&
-                (labelYear < value.year ||
-                  (labelYear === value.year && m < value.month))
-              ) {
-                css = "select";
-              }
-              if (atMinYear && m < ymArr[0].min) {
-                css = "disable";
-              }
-              if (atMaxYear && m > ymArr[yearMaxIdx].max) {
-                css = "disable";
-              }
-              if (otherValue) {
-                const y = otherValue.year;
-                const m = otherValue.month || 0;
-                const vy = labelYear;
-                const vm = i + 1;
-                if (
-                  y === vy &&
-                  m &&
-                  ((padIndex === 0 && vm > m) || (padIndex === 1 && vm < m))
-                ) {
-                  css = "disable";
-                } else if (
-                  (y > vy && padIndex === 1) ||
-                  (y < vy && padIndex === 0)
-                ) {
-                  css = "disable";
-                }
-              }
-              const clickHandler =
-                css !== "disable" ? handleClickMonth : undefined;
-              return (
-                <li
-                  key={i}
-                  className={classNames(styles.rmpBtn, styles[css])}
-                  data-id={`${padIndex}:${m}`}
-                  onClick={clickHandler}>
-                  {months.length > i ? months[i] : i}
-                </li>
-              );
-            })}
-          {!isMonthsMode &&
-            mapToArray(new Date(labelYear, labelMonth, 0).getDate(), (i) => {
-              let css = "";
-              const d = i + 1;
-              if (yearActive && labelMonth === value.month && d === value.day) {
-                css = "active";
-              }
-              // TODO: add max, min days
-              const clickHandler =
-                css !== "disable" ? handleClickDay : undefined;
-              return (
-                <li
-                  key={i}
-                  className={classNames(styles.rmpBtn, styles[css])}
-                  data-id={`${padIndex}:${labelMonth}:${d}`}
-                  onClick={clickHandler}>
-                  {d}
-                </li>
-              );
-            })}
-        </ul>
-      </div>
+      <DatePickerPad
+        key={padIndex}
+        padIndex={padIndex}
+        values={valuesState}
+        years={yearsState}
+        labelYear={labelYear}
+        labelMonth={labelMonth}
+        lang={lang}
+        yearIdx={yearIndexesState[padIndex]}
+        onMonthClick={handleClickMonth}
+        isMonthsMode={isMonthsMode}
+        onDayClick={handleClickDay}
+        onPrevYearClick={handlePrevYearClick}
+        onNextYearClick={handleNextYearClick}
+        onPrevMonthClick={handlePrevMonthClick}
+        onNextMonthClick={handleNextMonthClick}
+      />
     );
+  };
+
+  const renderPad = () => {
+    if (isRange > 1) {
+      setPads([padByIndex(0), padByIndex(1)]);
+    } else {
+      setPads([padByIndex(0)]);
+    }
   };
 
   const getValue = () => {
@@ -279,11 +112,6 @@ const DatePicker = ({
       _onDismiss();
       onClickAway && onClickAway(e);
     }
-  };
-
-  const _onDismiss = () => {
-    setShowedState(false);
-    onDismiss && onDismiss(getValue());
   };
 
   const handleClickMonth = useCallback((e) => {
@@ -309,19 +137,19 @@ const DatePicker = ({
     onChange(year, month, day, idx);
   }, []);
 
-  const handlePrevYearClick = (e) => {
+  const handlePrevYearClick = useCallback((e) => {
     const idx = parseInt(getDID(e), 10);
     if (yearIndexesState[idx] > 0) {
-      setYear(idx, -1);
+      _setYear(idx, -1);
     }
-  };
+  }, []);
 
-  const handleNextYearClick = (e) => {
+  const handleNextYearClick = useCallback((e) => {
     const idx = parseInt(getDID(e), 10);
     if (yearIndexesState[idx] < yearsState.length - 1) {
-      setYear(idx, 1);
+      _setYear(idx, 1);
     }
-  };
+  }, []);
 
   const handleNextMonthClick = (e) => {
     const idx = parseInt(getDID(e), 10);
@@ -334,7 +162,7 @@ const DatePicker = ({
     }
   };
 
-  const handlePrevMonthClick = (e) => {
+  const handlePrevMonthClick = useCallback((e) => {
     const idx = parseInt(getDID(e), 10);
     const labelMonth = labelMonthsState[idx];
     const nextMonth = labelMonth - 1;
@@ -343,9 +171,9 @@ const DatePicker = ({
       setLabelMonthsState(labelMonthsState);
       renderPad();
     }
-  };
+  }, []);
 
-  const setYear = (idx, step) => {
+  const _setYear = (idx, step) => {
     const yearIndex = (yearIndexesState[idx] += step);
     const labelYears = labelYearsState;
     const theYear = yearsState[yearIndex].year;
@@ -369,6 +197,11 @@ const DatePicker = ({
     }
     renderPad();
     onYearChange && onYearChange(theYear);
+  };
+
+  const _onDismiss = () => {
+    setShowedState(false);
+    onDismiss && onDismiss(getValue());
   };
 
   const _keyDown = useCallback((e) => {
@@ -396,7 +229,7 @@ const DatePicker = ({
             className={classNames(
               styles.rmpPopup,
               isMonthsMode && styles.monthsMode,
-              popupClass,
+              isRange && styles.range,
               styles[theme],
               showedState && styles.show
             )}>
